# docs/openapi/bravo-auth-service.yaml
openapi: 3.0.3
info:
  title: Bravo Secure — Auth Service
  version: 0.1.0
  description: |
    Authentication, user profile, session, Signal pre-key, TOTP, biometric
    and conversations REST surface for Bravo Secure.

    ## Getting started as a tester

    Follow these four steps and every protected endpoint will work from
    SwaggerHub's **Try it out** button:

    1. **Register.** Call `POST /auth/register` with a fresh `email`,
       `password`, `displayName`, and an E.164 `phoneE164`
       (e.g. `+15555550199`). Response is `{otpSentTo: "+15555550199"}`.
       The OTP is stubbed on staging — no real SMS is sent.
    2. **Verify registration.** Call `POST /auth/register/verify` with
       the same body **plus** `code: "1234"` (any 4–8 digit string
       passes on staging), a `deviceId` of your choice (e.g.
       `swaggerhub-1`), and `platform: "android"`. You get back
       `{accessToken, refreshToken, userId, ...}`.
    3. **Authorize.** Click the green **Authorize** button at the top
       of SwaggerHub, paste the `accessToken` into `BearerAuth`, and
       click **Authorize**. SwaggerHub will attach
       `Authorization: Bearer <accessToken>` to every subsequent call.
    4. **Exercise the API.** `GET /auth/me`, `GET /users/me`,
       `GET /conversations/mine`, etc. all now work. When the access
       token expires (default 15 min) exchange the refresh token via
       `POST /auth/refresh`.

    curl variant:

...

...
    fronted by Caddy + Let's Encrypt).

...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UnauthorizedError:
      type: object
      properties:
        statusCode: {type: integer, example: 401}
        error: {type: string, example: Unauthorized}
        message: {type: string, example: Unauthorized}
...
    KeyUploadBody:
      type: object
      required: [registrationId, identityKey, signedPrekeyId, signedPrekey, signedPrekeySig]
      properties:
        registrationId: {type: integer, minimum: 0, maximum: 16383}
        identityKey: {type: string, description: base64-encoded 32-byte Curve25519 public key}
        signedPrekeyId: {type: integer}
        signedPrekey: {type: string, description: base64-encoded 32-byte public key}
        signedPrekeySig:
          {type: string, description: base64-encoded 64-byte Ed25519 signature over signedPrekey}
        oneTimePrekeys:
          type: array
          maxItems: 100
          items:
            type: object
            required: [keyId, publicKey]
            properties:
              keyId: {type: integer}
              publicKey: {type: string, description: base64}
    KeyUploadResponse:
      type: object
      required: [ok, poolSize]
      properties:
        ok: {type: boolean, example: true}
        poolSize:
          {
            type: integer,
            example: 42,
            description: 'Remaining one-time pre-key count for this device',
          }
        oneTimeKeysStored:
          {
            type: integer,
            example: 2,
            description: 'How many OPKs from this request were newly inserted (source also returns this field)',
          }
    PreKeyBundle:
      type: object
      properties:
        userId: {type: string, format: uuid}
        registrationId: {type: integer}
        deviceId: {type: integer}
        identityKey: {type: string, description: base64}
        signedPreKey:
          type: object
          properties:
            keyId: {type: integer}
            publicKey: {type: string}
            signature: {type: string}
        preKey:
          type: object
          nullable: true
          properties:
            keyId: {type: integer}
            publicKey: {type: string}
...
  responses:
...
    Unauthorized:
      description: Missing / invalid bearer token
      content:
        application/json:
          schema: {$ref: '#/components/schemas/UnauthorizedError'}
...
paths:
...
  /auth/keys/upload:
    post:
      tags: [Keys]
      summary: Upload my Signal pre-key bundle
      description: |
        **Purpose.** Publish this device's Signal identity key,
        signed pre-key, signed pre-key signature, and a batch of
        one-time pre-keys into `public.signal_identities` /
        `public.signal_one_time_prekeys`. The signed pre-key is
        **Ed25519-signature-verified** against the identity key
        before it's stored (32-byte identity + 64-byte signature
        only — other sizes skip verification). OPK inserts are
        incremental and de-duped on `(user_id, key_id)`.

        NOTE: the yaml schema shows `signedPreKey.{keyId, publicKey,
        signature}` and `preKeys[].{keyId, publicKey}`. The actual
        DTO uses flat fields: `signedPrekeyId`, `signedPrekey`,
        `signedPrekeySig`, plus `oneTimePrekeys[].{keyId, publicKey}`.
        Use the flat field names in the curl below.

        **Auth.** Requires bearer access token.

        **Preconditions.** User must be fully registered. Base64-
        encoded keys must fit the size expectations (identity 32 B,
        sig 64 B). The response includes `poolSize`; once it drops
        below 10 the response also sets an `X-Pre-Key-Count` header
        nudging the client to replenish.

        **Example curl.**

            curl -X POST https://auth.94-136-184-52.sslip.io/auth/keys/upload \
              -H "Authorization: Bearer $ACCESS_TOKEN" \
              -H "Content-Type: application/json" \
              -d '{
                    "registrationId": 1234,
                    "identityKey": "<base64 32B>",
                    "signedPrekeyId": 1,
                    "signedPrekey": "<base64 32B>",
                    "signedPrekeySig": "<base64 64B>",
                    "oneTimePrekeys": [
                      {"keyId":100,"publicKey":"<base64 32B>"},
                      {"keyId":101,"publicKey":"<base64 32B>"}
                    ]
                  }'

        **Common failures.**
        - `400` — `signed_prekey_signature_invalid` (signature
          check failed); DTO validation (`registrationId` out of
          range, `oneTimePrekeys` > 100 entries, etc.).
        - `401` — bad bearer.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/KeyUploadBody'}
      responses:
        '200':
          description: Stored
          headers:
            X-Pre-Key-Count:
              description: |
                Emitted when the remaining one-time pre-key pool is below 10,
                nudging the client to replenish. Value equals `poolSize`.
              schema: {type: integer, example: 8}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KeyUploadResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/keys/{userId}:
    get:
      tags: [Keys]
      summary: Fetch a recipient's pre-key bundle
      description: |
        **Purpose.** Returns the target user's pre-key bundle so the
        caller can run Signal X3DH. **Consumes one** one-time pre-key
        atomically (SELECT + DELETE in the same SQL) so each OPK is
        handed out at most once. If the pool is empty the response
        comes back with `oneTimePrekey: null` (session creation
        falls back to just the signed pre-key). Response sets
        `X-Pre-Key-Count` header when fewer than 10 OPKs remain.

        **Auth.** Requires bearer access token.

        **Preconditions.** Target user must have previously called
        `POST /auth/keys/upload`. You can fetch your own bundle too —
        useful for smoke tests.

        **Example curl.**

            curl "https://auth.94-136-184-52.sslip.io/auth/keys/<userId>" \
              -H "Authorization: Bearer $ACCESS_TOKEN"

        **Common failures.**
        - `401` — bad bearer.
        - `404` — `no_keys_found`: the target has never uploaded a
          bundle.
      parameters:
        - in: path
          name: userId
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Bundle
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PreKeyBundle'}
        '404':
          description: User has no keys registered
        '401': {$ref: '#/components/responses/Unauthorized'}
...